+++ src/modules/db.js
@@ -1,4 +1,13 @@
 const fs = require("fs");
+const mysql = require("mysql");
+require("dotenv").config();
+
+var con = mysql.createConnection({
+  host: process.env.DATABASE_HOST,
+  user: process.env.DATABASE_USERNAME,
+  password: process.env.DATABASE_PASSWORD,
+  database: process.env.DATABASE_NAME
+});
 
 const checkDB = () =>{
   if(!fs.existsSync("./db.json")){