+++ src/modules/db.js
@@ -1,5 +1,6 @@
 const fs = require("fs");
 const mysql = require("mysql");
+const { lastGuess } = require("./embeds");
 require("dotenv").config();
 
 var con = mysql.createConnection({
@@ -16,48 +17,35 @@ const saveID = (userID, id) => {
   con.query(`INSERT INTO games VALUES (null, "${userID}", ${id})`);
 }
 const getUserGameID = (userID) => {
-  var file = fs.readFileSync('db.json').toString();
-  var db = JSON.parse(file);
-  db.games.forEach(element => {
-    for(const [key, val] of Object.entries(element)){
-      if(key == userID){
-        wordID = val.id;
-      }
-    }
+  return new Promise((resolve, reject) => {
+    con.query(`SELECT * FROM games WHERE userID="${userID}"`, (err, results) => {
+      if(results.length == 0) reject();
+      resolve(results[0].gameID);
+    });
   });
-  return wordID;
 }
 const addGuess = (userID, guess) => { //adds guess and returns all guesses
-  var file = fs.readFileSync('./db.json').toString();
-  var db = JSON.parse(file);
-  var newDB = {"games":[]};
-  var guesses = [];
-  var lastGuess = false;
-  // --- Update DB with new guess --- \\
-  db.games.forEach(element => {
-    for(const [key, val] of Object.entries(element)){
-      if(key == userID){
-        if(val.guesses == null){
-          val.guesses = [guess];
-        }else{
-          if(val.guesses.length == 5){
-            lastGuess = true;
-          }
-          val.guesses.push(guess);
-        }
-        guesses = val.guesses;
-        newDB.games.push({[userID]: {"id": wordID, "guesses": val.guesses}});
-      }else{
-        newDB.games.push({[key]: val});
-      }
-    }
+  return new Promise((resolve, reject) => {
+    con.query(`INSERT INTO guesses VALUES (null, "${userID}", "${guess}")`);
+    con.query(`SELECT guess FROM guesses WHERE userID="${userID}"`, (err, results) => {
+      var guesses = [];
+      results.forEach(result => {
+        guesses.push(result.guess);
+      });
+      resolve([guesses, results.length == 6]);
+    });
   });
-  fs.writeFileSync('./db.json', JSON.stringify(newDB, null, 2));
-  return [guesses, lastGuess];
+}
+const clearGameData = (userID) => {
+  return new Promise((resolve, reject) => {
+    con.query(`DELETE FROM games WHERE userID="${userID}"`);
+    con.query(`DELETE FROM guesses WHERE userID="${userID}"`);
+  });
 }
 
 module.exports = {
   saveID,
   getUserGameID,
-  addGuess
+  addGuess,
+  clearGameData
 }