+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog (Alpha Versions)
+## [Release Alpha 1.4.1](https://github.com/Nevah5/wordlebot/releases/tag/a1.4.1) - 08.02.2022
+### Added
+- Promise to saveid db function
+
 ## [Release Alpha 1.4](https://github.com/Nevah5/wordlebot/releases/tag/a1.4) - 08.02.2022
 ### Added
 - Mysql module and integration into db module